Seven split up
